+++ src/main/scala/org/podval/xml/CssClass.scala
@@ -1,0 +1,4 @@
+package org.podval.xml
+
+/** A CSS class token (not the HTML `class` attribute; that is [[HtmlClass]]). */
+open class CssClass(val name: String)