+++ src/main/scala/org/podval/xml/HtmlClass.scala
@@ -1,5 +1,0 @@
-package org.podval.xml
-
-object HtmlClass extends XmlAttribute("class")
-
-open class HtmlClass(val name: String)