+++ src/main/scala/org/podval/xml/XmlElement.scala
@@ -1,6 +1,9 @@
 package org.podval.xml
 
-open class XmlElement(val name: String)
+open class XmlElement(val expanded: XmlExpandedName):
+  def this(name: String) = this(XmlExpandedName.parse(name))
+
+  def name: String = expanded.qualifiedName
 
 object XmlElement:
   object A extends XmlElement("a")